Matched css with that of the designer
I made some changes on my pc as well, which were small and easy to implement. This method was a lot faster than running it through our designer at first.

We also made some changes together on my pc, which for that reason was never synchronized with what he had on his pc.

Next time, we use version control on our css as well.